+++ ebean-api/src/main/java/io/ebean/QueryBuilder.java
@@ -3,8 +3,6 @@ package io.ebean;
 import org.jspecify.annotations.Nullable;
 
 import jakarta.persistence.EntityNotFoundException;
-import javax.sql.DataSource;
-import java.sql.Connection;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Map;
@@ -13,8 +11,6 @@ import java.util.Set;
 import java.util.function.BooleanSupplier;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
-import java.util.function.Supplier;
-import java.util.stream.Stream;
 
 /**
  * Build and execute an ORM query.
@@ -22,7 +18,7 @@ import java.util.stream.Stream;
  * @param <SELF> The type of the builder
  * @param <T>    The entity bean type
  */
-public interface QueryBuilder<SELF extends QueryBuilder<SELF, T>, T> extends QueryBuilderProjection<SELF, T> {
+public interface QueryBuilder<SELF extends QueryBuilder<SELF, T>, T> extends QueryBuilderProjection<SELF, T>, StreamableQuery<SELF, T> {
 
   /**
    * Set root table alias.
@@ -149,46 +145,14 @@ public interface QueryBuilder<SELF extends QueryBuilder<SELF, T>, T> extends Que
   SELF copy();
 
   /**
-   * Execute the query using the given transaction.
-   */
-  SELF usingTransaction(Transaction transaction);
-
-  /**
    * Execute this query using immutable bean cache values for matching bean types.
    */
   SELF using(ImmutableBeanCache<?> beanCache);
 
   /**
-   * Execute the query using the given connection.
-   */
-  SELF usingConnection(Connection connection);
-
-  /**
    * Execute the query using the given database.
    */
   SELF usingDatabase(Database database);
-
-  /**
-   * Ensure that the master DataSource is used if there is a read only data source
-   * being used (that is using a read replica database potentially with replication lag).
-   * <p>
-   * When the database is configured with a read-only DataSource via
-   * say {@link io.ebean.config.DatabaseConfig#setReadOnlyDataSource(DataSource)} then
-   * by default when a query is run without an active transaction, it uses the read-only data
-   * source. We we use {@code usingMaster()} to instead ensure that the query is executed
-   * against the master data source.
-   */
-  default SELF usingMaster() {
-    return usingMaster(true);
-  }
-
-  /**
-   * Ensure the master DataSource is used when useMaster is true. Otherwise, the read only
-   * data source can be used if defined.
-   *
-   * @see #usingMaster()
-   */
-  SELF usingMaster(boolean useMaster);
 
   /**
    * Set the base table to use for this query.
@@ -681,51 +645,6 @@ public interface QueryBuilder<SELF extends QueryBuilder<SELF, T>, T> extends Que
   boolean exists();
 
   /**
-   * Execute the query returning either a single bean or null (if no matching
-   * bean is found).
-   * <p>
-   * If more than 1 row is found for this query then a PersistenceException is
-   * thrown.
-   * <p>
-   * This is useful when your predicates dictate that your query should only
-   * return 0 or 1 results.
-   * <p>
-   * <pre>{@code
-   *
-   * // assuming the sku of products is unique...
-   * Product product =
-   *     new QProduct()
-   *         .sku.equalTo("aa113")
-   *         .findOne();
-   * ...
-   * }</pre>
-   * <p>
-   * It is also useful with finding objects by their id when you want to specify
-   * further join information to optimise the query.
-   * <p>
-   * <pre>{@code
-   *
-   * // Fetch order 42 and additionally fetch join its order details...
-   * Order order =
-   *     new QOrder()
-   *         .fetch("details") // eagerly load the order details
-   *         .id.equalTo(42)
-   *         .findOne();
-   *
-   * // the order details were eagerly loaded
-   * List<OrderDetail> details = order.getDetails();
-   * ...
-   * }</pre>
-   */
-  @Nullable
-  T findOne();
-
-  /**
-   * Execute the query returning an optional bean.
-   */
-  Optional<T> findOneOrEmpty();
-
-  /**
    * Execute the query returning a single bean or throwing a {@link jakarta.persistence.EntityNotFoundException}
    * if there is no matching bean.
    * <p>
@@ -739,61 +658,11 @@ public interface QueryBuilder<SELF extends QueryBuilder<SELF, T>, T> extends Que
    * find-by-id query, or the single equality predicate when the query is filtered by what
    * looks like a natural/unique key, otherwise a generic "not found" message.
    */
+  @Override
   default T findOneOrThrow() {
     return findOneOrEmpty().orElseThrow(() ->
       new EntityNotFoundException(getBeanType().getSimpleName() + " not found"));
   }
-
-  /**
-   * Execute the query returning a single bean or throwing the exception produced by the
-   * given supplier if there is no matching bean.
-   * <pre>{@code
-   *   Customer customer = query
-   *     .findOneOrThrow(() -> new NotFoundException("Customer not found for id: " + id));
-   * }</pre>
-   */
-  default T findOneOrThrow(Supplier<? extends RuntimeException> exceptionSupplier) {
-    return findOneOrEmpty().orElseThrow(exceptionSupplier);
-  }
-
-  /**
-   * Execute the query returning the list of objects.
-   * <p>
-   * This query will execute against the EbeanServer that was used to create it.
-   * <p>
-   * <pre>{@code
-   *
-   * List<Customer> customers =
-   *     new QCustomer()
-   *       .name.ilike("rob%")
-   *       .findList();
-   *
-   * }</pre>
-   *
-   * @see Query#findList()
-   */
-  List<T> findList();
-
-  /**
-   * Execute the query returning the result as a Stream.
-   * <p>
-   * Note that this can support very large queries iterating
-   * any number of results. To do so internally it can use
-   * multiple persistence contexts.
-   * </p>
-   * <pre>{@code
-   *
-   *  // use try with resources to ensure Stream is closed
-   *
-   *  try (Stream<Customer> stream = query.findStream()) {
-   *    stream
-   *    .map(...)
-   *    .collect(...);
-   *  }
-   *
-   * }</pre>
-   */
-  Stream<T> findStream();
 
   /**
    * Execute the query returning the set of objects.
@@ -1080,34 +949,5 @@ public interface QueryBuilder<SELF extends QueryBuilder<SELF, T>, T> extends Que
    * @return a Future object for the map result of the query
    */
   <K> FutureMap<K,T> findFutureMap();
-
-  /**
-   * Return a PagedList for this query using firstRow and maxRows.
-   * <p>
-   * The benefit of using this over findList() is that it provides functionality to get the
-   * total row count etc.
-   * <p>
-   * If maxRows is not set on the query prior to calling findPagedList() then a
-   * PersistenceException is thrown.
-   * <p>
-   * <pre>{@code
-   *
-   *  PagedList<Order> pagedList =
-   *    new QOrder()
-   *       .setFirstRow(50)
-   *       .setMaxRows(20)
-   *       .findPagedList();
-   *
-   *       // fetch the total row count in the background
-   *       pagedList.loadRowCount();
-   *
-   *       List<Order> orders = pagedList.getList();
-   *       int totalRowCount = pagedList.getTotalRowCount();
-   *
-   * }</pre>
-   *
-   * @return The PagedList
-   */
-  PagedList<T> findPagedList();
 
 }
